text-analysis-service/functions/text_analysis/requirements.txt
numpy>=2.3.5
scikit-learn>=1.8.0
pydantic>=2.12.5
# transformers>=4.30.0  # Commented out for now to avoid large downloads
vaderSentiment>=3.3.2
torch>=2.0.0
sentence-transformers>=2.2.2

# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
